# DateTimeService/wwwroot/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "DateTime Service",
    "description": "Сервис получения дат доступности и интервалов доставки/самовывоза. Версия 2.0 переписана на .Net и прямые обращения к SQL.",
    "termsOfService": "",
    "version": "2.0",
    "contact": {
      "name": "Андрей Бородавко",
      "email": "a.borodavko@21vek.by"
    }
  },
...
  "servers": [
    {
      "url": "https://courier.21vek.by:739/NetMS"
    }
  ],
...
  "paths": {
    "/api/DateTime/IntervalList": {
      "post": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "DateTime"
        ],
        "summary": "Получить интервалы доставки",
        "description": "Получает список доступных интервалов доставки",
        "requestBody": {
          "description": "содержимое отправляемого запроса",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/payload_interval"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "запрос выполнен",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "properties": {
                          "begin": {
                            "type": "string",
                            "format": "date-time",
                            "description": "Начало интервала"
                          },
                          "end": {
                            "type": "string",
                            "format": "date-time",
                            "description": "Конец интервала"
                          }
                        }
                      },
                      "example": [
                        {
                          "begin": "2021-03-23T19:00:00",
                          "end": "2021-03-23T21:00:00"
                        },
                        {
                          "begin": "2021-03-23T21:00:00",
                          "end": "2021-03-23T23:00:00"
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Не аутентифицирован"
          },
          "403": {
            "description": "Отсутствуют права на данный эндпоинт"
          },
          "500": {
            "description": "внутренняя ошибка сервера",
            "content": {}
          }
        },
        "x-codegen-request-body-name": "body"
      }
    },
    "/api/DateTime/AvailableDate": {
      "post": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "DateTime"
        ],
        "summary": "Получить даты доступности",
        "description": "Получает ближайшие доступные даты доставки и самовывоза по каждому товару",
        "requestBody": {
          "description": "содержимое отправляемого запроса",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/payload_availabledate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "запрос выполнен",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "object",
                        "description": "Ключ - артикул",
                        "properties": {
                          "code": {
                            "type": "string",
                            "description": "Артикул"
                          },
                          "courier": {
                            "type": "string",
                            "format": "date-time",
                            "description": "Дата курьерской доставки"
                          },
                          "self": {
                            "type": "string",
                            "format": "date-time",
                            "description": "Дата самовывоза"
                          }
                        }
                      }
                    }
                  }
                },
                "example": {
                  "data": {
                    "101937": {
                      "code": "101937",
                      "courier": "2021-03-23T13:00:00",
                      "self": "3999-11-11T00:00:00"
                    },
                    "1020938": {
                      "code": "1020938",
                      "courier": "2021-03-24T19:00:00",
                      "self": "3999-11-11T00:00:00"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Не аутентифицирован"
          },
          "403": {
            "description": "Отсутствуют права на данный эндпоинт"
          },
          "500": {
            "description": "внутренняя ошибка сервера",
            "content": {}
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "payload_interval": {
        "properties": {
          "address_id": {
            "type": "string",
            "description": "Код СОАТО адреса",
            "example": "47175"
          },
          "delivery_type": {
            "type": "string",
            "description": "Вид доставки - самовывоз (self) или курьер (courier). Если не указан, то считается как доставка",
            "example": "self"
          },
          "floor": {
            "type": "number",
            "description": "Этаж",
            "example": 8
          },
          "payment": {
            "type": "string",
            "description": "Тип оплаты: partly_pay, legal_entity, other",
            "example": "other"
          },
          "order_items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/orderitem"
            },
            "description": "Список товаров",
            "example": [
              {
                "code": "631176",
                "quantity": 1
              },
              {
                "code": "467878",
                "quantity": 2
              },
              {
                "code": "609204",
                "quantity": 1
              },
              {
                "code": "6082000",
                "quantity": 6
              }
            ]
          }
        }
      },
      "payload_availabledate": {
        "properties": {
          "city_id": {
            "type": "string",
            "description": "Код города",
            "example": "17030"
          },
          "delivery_types": {
            "type": "array",
            "description": "Вид доставки - самовывоз (self) или курьер (courier)",
            "items": {
              "type": "string"
            },
            "example": [
              "self",
              "courier"
            ]
          },
          "codes": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Список артикулов",
            "example": [
              "358649",
              "424941",
              "1020938",
              "5962720",
              "6167903",
              "380386",
              "101937"
            ]
          }
        }
      },
      "orderitem": {
        "properties": {
          "code": {
            "type": "string",
            "description": "Артикул номенклатуры"
          },
          "quantity": {
            "type": "number",
            "description": "Количество"
          }
        },
        "required": [
          "code",
          "quantity"
        ]
      },
...
    }
  }
}